requirements.txt
streamlit
joblib
numpy
matplotlib
pandas
scikit-learn
scipy
opencv-python-headless
Pillow
tslearn